Move serialization into the storage component.
The serialization and deserialization of the data used to be in the Rollout
component, which meant that the storage API dealt exclusively with string
values. This sort of locked us into a string-based persistence implementation.
By moving the serialization into the storage, it opens us up to having other
kinds of persistence (such as MySQL) that may deal with more of a relational
style of persistence.